move settings switches to settings folder
