+++ app/settings/SavingIndicatorSwitch.test.tsx
@@ -1,5 +1,5 @@
 import { expect, it, vi } from 'vitest'
-import { render, screen } from '../lib/test/rtl'
+import { render, screen } from '../../lib/test/rtl'
 import SavingIndicatorSwitch from './SavingIndicatorSwitch'
 
 it('reads value from local storage', async () => {
